+++ src/background/state/application/toolbar_items.rs
@@ -29,6 +29,7 @@ impl FullState {
             center: vec![ToolbarItem2::Plugin("@default/date".into())],
             right: vec![
                 ToolbarItem2::Plugin("@default/system-tray".into()),
+                ToolbarItem2::Plugin("@default/keyboard".into()),
                 ToolbarItem2::Plugin("@default/bluetooth".into()),
                 ToolbarItem2::Plugin("@default/network".into()),
                 ToolbarItem2::Plugin("@default/media".into()),